Move custom events to ui folder to simplify build.
