+++ src/ui/BalancePanel.cpp
@@ -3,7 +3,7 @@
 #include "api/Api.hpp"
 #include "ui/BalancePanel.hpp"
 #include "ui/StaticCurrency.hpp"
-#include "util/event.hpp"
+#include "ui/event.hpp"
 
 // constructor ////////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
