+++ src/ui/LoginPanel.cpp
@@ -2,7 +2,7 @@
 
 #include "api/Api.hpp"
 #include "ui/LoginPanel.hpp"
-#include "util/event.hpp"
+#include "ui/event.hpp"
 
 // constructor ////////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
