+++ src/ui/MainFrame.cpp
@@ -5,7 +5,7 @@
 #include "ui/MainFrame.hpp"
 #include "ui/LoginPanel.hpp"
 #include "ui/PortfolioPanel.hpp"
-#include "util/event.hpp"
+#include "ui/event.hpp"
 
 // constructor ////////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
