+++ src/ui/PortfolioPanel.cpp
@@ -5,7 +5,7 @@
 #include "ui/PortfolioPanel.hpp"
 #include "ui/EventPositionPanel.hpp"
 #include "ui/MarketPositionPanel.hpp"
-#include "util/event.hpp"
+#include "ui/event.hpp"
 
 // constructor ////////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
